requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
supabase==2.0.2
redis==5.0.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aioredis==2.0.1
python-multipart==0.0.6
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.23.2